backend/backend_requirements.txt
fastapi
numpy
pillow
python-multipart
uvicorn
transformers